refactor: utilize bubbling up `<title>` element
I was reading the RRv7 docs and found that it mentions that React 19+ actually bubbles up `<meta>` and `<title>` elements to the `<head>` element.

Thus, RRv7's docs don't outright state it, but the old `meta` function export is basically deprecated. This version Ypiretis was ported from a Remix v2 testbed where `meta` functions were the only way.

This is fine. It helps to simplify a lot.

https://reactrouter.com/start/framework/route-module#meta