refactor: entities now have auto increment lookup ids
Display entities are not associated with any particular user or anything. Like how there is `IPresenterUser.user` and `IAttendeeUser.user` via `IUser.user`.

And a presenter can have N amount of display entities at any given time. So, we need a way to reference them in the future. Like say, a presenter wants to "kick" the display via their presenter view. Rather than closing the window.

Why a presenter might want to do that? I don't know. Maybe we have a situation where there's a display entity in the back of a classroom running on a separate computer from the display entity in the front of the classroom running off of the presenter's host computer. In any case, we need some primitive value to reference display entities.