feat: initial websocket support for RRv7
I was running into issues of RRv7 (or Hono via Bun) not handling SSE timeouts well. Basically, after five seconds, the SSE connection would terminate.

Also, Hono via Bun does not have native HTTP2 support. Meaning, we would run into the limitations imposed on HTTP 1 SSE in Browser.

"Warning: When not used over HTTP/2, SSE suffers from a limitation to the maximum number of open connections, which can be especially painful when opening multiple tabs, as the limit is per browser and is set to a very low number (6). The issue has been marked as "Won't fix" in Chrome and Firefox. This limit is per browser + domain, which means that you can open 6 SSE connections across all of the tabs to www.example1.com and another 6 SSE connections to www.example2.com (per Stack Overflow). When using HTTP/2, the maximum number of simultaneous HTTP streams is negotiated between the server and the client (defaults to 100)."

**SOURCE:** https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events/Using_server-sent_events