refactor: now using flash cookies as a bearer source
Browsers throw an error during `new WebSocket` _if_ you use a non-standard value for the protocols. So.... next way to hold bearer values w/o putting them in in-secure query params would be cookies.

The still-secure alternative here would be to use the consent token to generate one-time connection tokens that can be used to initiated web socket connections. But that requires more server state. So, this is the simplier way to go while remaining secure.