feat: updated article edited flag to be accurate down to the minute instead of milliseconds
The public would not care about what second nor millisecond when it was published. And we do not display that information anyway.

The main reason for doing this is because the update at timestamp hook runs after every time a change has occured. So, if the user switches the article from draft -> published, then the updated at timestamp would be _after_ the publish at timestamp. Which is not at all the intention of what the two columns represent.

We could handle the update at timestamp at the endpoints that set the article state instead of leaving it be automagical. That would sidestep this issue completly. But then we might let it slip somewhere when the article data was last updated.