refactor: use async button update workflow like content updates for updating the published at date
Because of how the browser's native datetime control works the input updates it value whenever _any_ datetime component is changed by the user. Rather than waiting for the user to blur the input / commit in some way.

This lead to the user experience being you type in the hour datetime component -> update is submitted automagically and input blurs -> user has to refocus input -> user types minute datetime component -> ... -> and repeat. It was a very disjointed UX.