fix: unreachable code error
This was an interesting issue. In VS Code, the built-in TypeScript analyzer presented the (intended) unreachable as a warning. But when running the TypeScript compiler it was presented as an error. In any case, moving the fallback return statement into the default case silenced the error and warning. And since it is no longer a warning I went ahead and added a comment.