removal: MDXEditor's [Markdown shortcuts feature](https://mdxeditor.dev/editor/docs/markdown-shortcuts)
I loved this feature but the plugin tried to use the Prism code edtior global (why?????? it's a shortcuts plugin!) which caused runtime errors. And the work around to actually make it work would have caused too much complication. Because I tried to just import `prismjs` and add it to `globalThis.Prism` but `rolldown-vite` removed it in production builds. And I would have had to dynamically structured the `MarkdownEditor` component around runtime imports with suspense... and that is just not worth it.