fix: attendees not being recognized as approved after a presenter approves their's join
I completely forgot that getters and setters work on a prototype-level. So, when I was spreading the base object into the child object the getters were being evaluated for their values at the time and those values were being spread into the child objects. Because of that, that meant the private values in the base enclosures weren't being replicated down the inheritance chain when updated.