refactor: use Temporal API instead of Date API
We are now shipping a ~47 KB (min+gzip) polyfill of the Temporal API on all pages that use anything calendar related. Yaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaay! Not ideal is the least of it! On the other hand... providing users with accurate dates and times is more important. Since I needed to handle timezone stuff I could not really use the Date API as there is no way to manually set timezones like there is with the Temporal API.